Combat flags for instances

+++ src/creature_instances.h
@@ -100,7 +100,7 @@ struct InstanceInfo {
     long reset_time;
     long fp_reset_time;
     unsigned char graphics_idx;
-    short flags;
+    unsigned long flags;
     short force_visibility;
     unsigned char primary_target;
     unsigned char func_idx;
